Transfer access token grant flow
Transferred the implementation flow of each
access token grant flow to respective rs file
