+++ src/auth_code_grant.rs
@@ -1,16 +1,21 @@
 // Standard libraries
+use std::io::{BufRead, BufReader, Write};
+use std::net::TcpListener;
+use std::path::PathBuf;
 use std::{future::Future, path::Path};
 
 // 3rd party crates
 use async_trait::async_trait;
-use oauth2::AuthorizationCode;
+use directories::UserDirs;
 use oauth2::{
     basic::BasicClient, url::Url, AuthUrl, ClientId, ClientSecret, CsrfToken, HttpRequest,
     HttpResponse, RedirectUrl, Scope, TokenUrl,
 };
+use oauth2::{AccessToken, AuthorizationCode};
 
 // My crates
 use crate::error::{ErrorCodes, OAuth2Error, OAuth2Result};
+use crate::http_client::async_http_client;
 use crate::TokenKeeper;
 
 #[async_trait]
@@ -162,4 +167,95 @@ impl AuthCodeGrant {
         )
         .set_auth_type(oauth2::AuthType::RequestBody))
     }
+}
+
+pub async fn auth_code_grant(
+    client_id: &str,
+    client_secret: Option<ClientSecret>,
+    sender_email: &str,
+) -> OAuth2Result<AccessToken> {
+    let auth_code_grant = AuthCodeGrant::new(
+        ClientId::new(client_id.to_string()),
+        client_secret,
+        AuthUrl::new("https://login.microsoftonline.com/common/oauth2/v2.0/authorize".to_string())?,
+        TokenUrl::new("https://login.microsoftonline.com/common/oauth2/v2.0/token".to_string())?,
+    );
+    let scopes = vec![
+        Scope::new("offline_access".to_string()),
+        Scope::new("SMTP.Send".to_string()),
+    ];
+    let directory = UserDirs::new().unwrap();
+    let mut directory = directory.home_dir().to_owned();
+
+    directory = directory.join("token");
+
+    let token_file = PathBuf::from(format!("{}_auth_code_grant.json", sender_email));
+    let mut token_keeper = TokenKeeper::new(directory.to_path_buf());
+
+    // If there is no exsting token, get it from the cloud
+    if let Err(_err) = token_keeper.read(&token_file) {
+        let (authorize_url, _csrf_state) =
+            auth_code_grant.generate_authorization_url(scopes).await?;
+        log::info!(
+            "Open this URL in your browser:\n{}\n",
+            authorize_url.to_string()
+        );
+
+        let listener = TcpListener::bind("127.0.0.1:8080").unwrap();
+        if let Some(mut stream) = listener.incoming().flatten().next() {
+            let code;
+            let _state;
+            {
+                let mut reader = BufReader::new(&stream);
+
+                let mut request_line = String::new();
+                reader.read_line(&mut request_line).unwrap();
+
+                let redirect_url = request_line.split_whitespace().nth(1).unwrap();
+                let url = Url::parse(&("http://localhost".to_string() + redirect_url)).unwrap();
+
+                let code_pair = url
+                    .query_pairs()
+                    .find(|pair| {
+                        let (key, _) = pair;
+                        key == "code"
+                    })
+                    .unwrap();
+
+                let (_, value) = code_pair;
+                code = AuthorizationCode::new(value.into_owned());
+
+                let state_pair = url
+                    .query_pairs()
+                    .find(|pair| {
+                        let (key, _) = pair;
+                        key == "state"
+                    })
+                    .unwrap();
+
+                let (_, value) = state_pair;
+                _state = CsrfToken::new(value.into_owned());
+            }
+
+            let message = "Go back to your terminal :)";
+            let response = format!(
+                "HTTP/1.1 200 OK\r\ncontent-length: {}\r\n\r\n{}",
+                message.len(),
+                message
+            );
+            stream.write_all(response.as_bytes()).unwrap();
+
+            // Exchange the code with a token.
+            token_keeper = auth_code_grant
+                .exchange_auth_code(&directory, &token_file, code, async_http_client)
+                .await?;
+
+            // The server will terminate itself after collecting the first code.
+        }
+    } else {
+        token_keeper = auth_code_grant
+            .get_access_token(&directory, &token_file, async_http_client)
+            .await?;
+    }
+    Ok(token_keeper.access_token)
 }
